Conveyorize!
1. Add a GitHub Actions CI job to compile the JS.

The GitHub Desktop build system does a lot of customizations
based on what the current OS is. Although it could be fixed
to support cross-building, it's easier to just use Actions to
build the out/ directory on each target platform.

This requires some small adjustments to build.ts so we can
produce the input files without running Electron Packager.

2. Add conveyor.conf

3. Update the README

4. Adjust the startup code to be more flexible about URL
   parsing. This is needed because the way URL open requests
   are received with Conveyor is different.